### 2017-04-16 #240 @whedon
[PRE REVIEW]: Pandarus: GIS toolkit for regionalized life cycle assessment
**Submitting author:** @cmutel (<a href="http://orcid.org/0000-0002-7898-9862">Chris Mutel</a>)
**Repository:** <a href="https://github.com/cmutel/pandarus" target ="_blank">https://github.com/cmutel/pandarus</a>
**Version:** 1.0.1
