### 2017-04-16 #240 @cmutel
@jsta, @nuest, @sgrieve or especially @hugoledoux - would any of you have time for this? Nothing too revolutionary, buts fills a specific needs in a research community that is not very geographically-aware.
