### 2017-04-17 cc49617
Fix license text
