+++ LICENSE.txt
@@ -1,16 +1,16 @@
-Copyright (c) 2016, Chris Mutel and ETH Zürich
-All rights reserved.
+Copyright (c) 2016, Chris Mutel and ETH Zurich.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
 
 Redistributions of source code must retain the above copyright notice, this
-list of conditions and the following disclaimer. Redistributions in binary
-form must reproduce the above copyright notice, this list of conditions and the
-following disclaimer in the documentation and/or other materials provided
-with the distribution.
+list of conditions and the following disclaimer. 
 
-Neither the name of ETH Zürich nor the names of its contributors may be used
+Redistributions in binary form must reproduce the above copyright notice, 
+this list of conditions and the following disclaimer in the documentation 
+and/or other materials provided with the distribution.
+
+Neither the name of ETH Zurich nor the names of its contributors may be used
 to endorse or promote products derived from this software without specific
 prior written permission.
 
