+++ pandarus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = (1, 0, 1)
+__version__ = (1, 0, 2)
 
 __all__ = (
     "Map",
