### 2017-04-18 317caa8
Small setup changes
