+++ setup.py
@@ -17,7 +17,7 @@ setup(
     author="Chris Mutel",
     author_email="cmutel@gmail.com",
     classifiers=[
-        'Development Status :: 5 - Production/Stabl',
+        'Development Status :: 5 - Production/Stable',
         'Intended Audience :: End Users/Desktop',
         'Intended Audience :: Developers',
         'Intended Audience :: Science/Research',
