### 2017-04-19 #244 @arfon
Hi @hugoledoux - please work your way through the review checklist above to complete your review (you should be able to check items off the list as you verify them). Please note we have reviewer guidelines here http://joss.theoj.org/about#reviewer_guidelines for you but if you have any questions for me or the author please ask here.
