### 2017-04-19 #244 @cmutel
Software is uploaded to [Zenodo](https://zenodo.org/record/556151) with the DOI 10.5281/zenodo.556151.
