### 2017-05-02 #244 @hugoledoux
I'm one of those who still use Python 2.7... but it can't be installed as Python >3.0 is required, right?
