### 2017-05-02 #244 @hugoledoux
In general nice and well-documented library, and I'm sure it's really useful for people that are non-expert at GIS.

I have the following remarks below. I tried it a bit but to continue I'd really like to have one full example, as explained below. I know I could figure things out, but as a user, I like to just see quickly what it does, and not having to build my own script from scratch.

## version

The version is not 1.0.1 but 1.0.2. I assume this ain't a problem.

# installation

Only Python >3.0 is supported, which is not mentioned and the setup.py (or used with pip) doesn't complain, but an error is thrown because the open() function of Python 2.7 for instance doesn't have the `encoding` option.

Installing with conda works fine, if Python 3.6 is chosen.

## Software paper

I'd say you 'overlay' 2 GIS datasets, and not only 'match'. This is in my opinion the more technical term used in GIS.

"Calculate the areas left out (present in one input file but not the other) from the aforementioned intersections calculation"
--> a Boolean difference is thus performed could also be stated

## Example usage

As a GIS person, I'd like to have one full example (a full python script) that reads a shapefile (or gpkg or geojson), does the intersection with another file (eg the one in your first example "Intersecting two vector datasets") and outputs to a another shapefile. I looked but couldn't find it, only bits of code (which are useful, don't get me wrong!) and the tests. The test datasets are already included, so why not have 2-3 full examples?
