### 2017-05-02 0861cd6
Update readme and docs for review comments
