+++ docs/index.rst
@@ -3,10 +3,10 @@ Pandarus
 
 Pandarus is a GIS software toolkit for regionalized life cycle assessment (LCA). It is designed to work with `brightway LCA framework <https://brightwaylca.org>`__, `brightway2-regional <https://bitbucket.org/cmutel/brightway2-regional>`__, and `Constructive Geometries <https://bitbucket.org/cmutel/constructive-geometries>`__. A separate library, `pandarus-remote <https://github.com/cmutel/pandarus_remote>`__, provides a web API to run Pandarus on a server.
 
-In the context of life cycle assessment, regionalization means the introduction of detailed spatial information for inventory activities and impact assessment characterization maps. As these will have different spatial scales, GIS functionality is required to match these two maps. Pandarus can do the following:
+In the context of life cycle assessment, regionalization means the introduction of detailed spatial information for inventory activities and impact assessment characterization maps. As these will have different spatial scales, GIS functionality is required to overlay these two maps. Pandarus can do the following:
 
-* Match two vector datasets, calculating the areas of each combination of features using the Mollweide projection.
-* Calculate the areas left out (present in one input file but not the other) from the aforementioned intersections calculation.
+* Overlay two vector datasets, calculating the areas of each combination of features using the Mollweide projection.
+* Calculate the area of the geometric difference (the areas present in one input file but not in the other) of one vector dataset with another vector dataset.
 * Calculate statistics such as min, mean, and max when overlaying a raster dataset with a vector dataset.
 * Normalize raster datasets, including use of compatible `nodata` values
 * Vectorization of raster datasets
@@ -119,7 +119,7 @@ FAQ
 Why the name Pandarus?
 ----------------------
 
-The software matches two different maps against each other, and `Pandarus was a bit of a matchmaker himself <http://en.wikipedia.org/wiki/Pandarus>`_. Plus, ancient names are 172% more science-y.
+The software overlays (matches) two different maps, and `Pandarus was a bit of a matchmaker himself <http://en.wikipedia.org/wiki/Pandarus>`_. Plus, ancient names are 172% more science-y.
 
 Installation
 ============
@@ -138,7 +138,7 @@ However, it is easy to run into errors if geospatial libraries like fiona, raste
     activate pandarus  # Could also be source activate pandarus
     conda install -c conda-forge -c cmutel pandarus
 
-Pandarus source code is on `GitHub <https://github.com/cmutel/pandarus>`__.
+Pandarus is only compatible with Python >= 3.5. Source code is on `GitHub <https://github.com/cmutel/pandarus>`__.
 
 Requirements
 ------------
