### 2017-05-02 2f02b92
Port changes to paper.md file
