### 2017-05-02 #244 @cmutel
@hugoledoux Updated the readme and docs:

- Updated language (please let me know if there are other suggestions!)
- Made usage example easier to find
- Added notes on Python 3 requirement

Please feel free to make technical criticisms (esp. using the Mollweide as a default projection) as well!
