### 2017-05-03 #244 @hugoledoux
@cmutel

## Example usage

I had seen that `ipynb` file, but when I opened it I had no idea what it was. I am a casual user of Python, and do not use advanced features... Am I the "average user" for your library? I wouldn't know, but I think having an example directly runnable with python would surely help.

I tried converting it to a `py` file, but ran into issues while running... macOS complained about Python not being installed as a framework.
I also tried the Jupyter notebook thing but I can't run it without errors, and I don't know what I am doing wrong here. The package pandarus is not recognised, while I started jupyter from the conda env... Any idea what I might be overlooking?

## Projection and CRS

    Because Pandarus was designed for global data sets, the Mollweide projection is used as the default equal-area projection for calculating areas (in square meters). Although no projection is perfect, the Mollweide has been found to be a reasonable compromise

--> I agree with the projection by default. However, do you always reproject? I'd would argue that the default should be no reprojection as users often have their data in the same (local) CRS, no? In the Netherlands it's certainly the case. I would argue that you should automatically reproject to Mollweide only when:

1. the input datasets are in a different CRS
2. they are both in WGS84 lat-long (which will be often the case I reckon for your users?)

## Writing output files

I watched your video (it's nice and explains well what you've done) and you state that you can export files in different GIS formats (and not just Python objects), which I would expect since you use fiona/rasterio. But it seems that in the example given the results are only outputted in json, is that correct? You have one function `pandarus.filesystem.json_exporter()`, and `pandarus.maps.Map()` only wraps `open`, not `write`.

And why json and not geojson? Your downstream software use this? I'd argue that having a SHP or geojson output would be super useful, and easy to do.
