### 2017-05-04 4dbf364
Add usage example as a python file
