### 2017-05-04 fe01a07
Clarify that geojson files are created
