### 2017-05-04 #244 @cmutel
@hugoledoux Thanks for your detailed questions!

## Example usage

Not sure what went wrong in your installation, I have successfully tested the [installation instructions](http://pandarus.readthedocs.io/?badge=latest#installation) on OS X and Windows. Maybe there are path problems due to multiple Python installations? I normally don't make miniconda the default Python, and so have to do things a but more manually, e.g. `source ~/miniconda3/bin/activate pandarus-test`/

I added the example usage as a [plain Python file](https://github.com/cmutel/pandarus/commit/4dbf364177da4f06b366a68b1458ec55c8c1895f).

## Projection and CRS

I used to skip projecting if both datasets had the same CRS, but this was buggy - it is hard to compare CRS strings/objects, at least in Python. I ended up always projecting to WGS 84 and then Mollweide, as this was guaranteed to work, and almost all spatial datasets should be global (we are modelling/assessing global supply chains).

## Writing output files

The output data here would almost always be consumed by downstream software which only cares about the numeric values. Nevertheless, I modified the paper description and example usage to be more explicit that geojson is also output where it would be useful, e.g. for all possible intersections of one spatial file with another.
