### 2017-05-04 #244 @hugoledoux
@cmutel

No idea about the Jupyter thing, but that's obviously on my side. The added py file works fine, including the output to geojson.

Projection: as a GIS person, I still think that projecting to something else is not optimal, but I also understand that your users are probably not even aware of CRS. So your solution is probably the best.

As far as I am concerned, this is it. Your library is nice, and it's nice to see the efforts you've put in making it and document it. It makes a bit ashamed of my own projects, so that's a good outcome of this review for me: going back to work and better document my own stuff.

@arfon it's my first review, so let me know if that was "what is expected".
