### 2017-05-04 #244 @arfon
> @arfon it's my first review, so let me know if that was "what is expected".

@hugoledoux - I think you've done a great job here. Can you confirm that @cmutel has addressed your feedback? If so, I think we should be good to accept here.
