### 2017-05-04 #244 @hugoledoux
He has addressed all issues indeed. I'd accept.
