### 2017-05-04 #244 @arfon
> He has addressed all issues indeed. I'd accept.

👍 thanks for confirming.
