+++ setup.py
@@ -13,7 +13,7 @@ requirements = [
 ]
 
 setup(
-    version="1.0.2",
+    version="1.0.3",
     author="Chris Mutel",
     author_email="cmutel@gmail.com",
     classifiers=[
