### 2017-05-04 #244 @cmutel
@arfon New version with revision sugested by Hugo uploaded to [Zenodo](https://zenodo.org/record/571547) with DOI 10.5281/zenodo.571547. Note version update to 1.0.4.

Thanks everyone!
