### 2017-05-04 b0badc4
1.0.4. Include license file.
