+++ MANIFEST.in
@@ -1,3 +1,4 @@
+include LICENSE
 include *.txt
 include README.md
 include pandarus/*.py
